fixed errorcase3

+++ roqoqo-qryd/tests/integration/api_backend.rs
@@ -826,48 +826,50 @@ fn api_backend_errorcase_const() {
 /// Test error cases. Case 3: invalid API TOKEN
 #[test]
 fn api_backend_errorcase3() {
-    let number_qubits = 6;
-    let device = QrydEmuSquareDevice::new(Some(2), Some(0.23));
-    let qryd_device: QRydAPIDevice = QRydAPIDevice::from(&device);
+    if !env::var("QRYD_API_TOKEN").is_ok() {
+        let number_qubits = 6;
+        let device = QrydEmuSquareDevice::new(Some(2), Some(0.23));
+        let qryd_device: QRydAPIDevice = QRydAPIDevice::from(&device);
 
-    let api_backend_err = APIBackend::new(qryd_device.clone(), None, None, None);
-    assert!(api_backend_err.is_err());
-    assert_eq!(
-        api_backend_err.unwrap_err(),
-        RoqoqoBackendError::MissingAuthentification {
-            msg: "QRYD access token is missing".to_string()
-        }
-    );
+        let api_backend_err = APIBackend::new(qryd_device.clone(), None, None, None);
+        assert!(api_backend_err.is_err());
+        assert_eq!(
+            api_backend_err.unwrap_err(),
+            RoqoqoBackendError::MissingAuthentification {
+                msg: "QRYD access token is missing".to_string()
+            }
+        );
 
-    let api_backend_new =
-        APIBackend::new(qryd_device, Some("DummyString".to_string()), None, None).unwrap();
+        let api_backend_new =
+            APIBackend::new(qryd_device, Some("DummyString".to_string()), None, None).unwrap();
 
-    let mut circuit = Circuit::new();
-    circuit += operations::DefinitionBit::new("ro".to_string(), number_qubits, true);
-    circuit += operations::RotateX::new(0, std::f64::consts::PI.into());
-    circuit += operations::RotateX::new(4, std::f64::consts::FRAC_PI_2.into());
-    circuit += operations::MeasureQubit::new(0, "ro".to_string(), 0);
-    circuit += operations::PragmaSetNumberOfMeasurements::new(10, "ro".to_string());
-    let measurement = ClassicalRegister {
-        constant_circuit: None,
-        circuits: vec![circuit.clone()],
-    };
-    let program = QuantumProgram::ClassicalRegister {
-        measurement,
-        input_parameter_names: vec![],
-    };
-    let job_loc = api_backend_new.post_job(program);
-    assert!(job_loc.is_err());
+        let mut circuit = Circuit::new();
+        circuit += operations::DefinitionBit::new("ro".to_string(), number_qubits, true);
+        circuit += operations::RotateX::new(0, std::f64::consts::PI.into());
+        circuit += operations::RotateX::new(4, std::f64::consts::FRAC_PI_2.into());
+        circuit += operations::MeasureQubit::new(0, "ro".to_string(), 0);
+        circuit += operations::PragmaSetNumberOfMeasurements::new(10, "ro".to_string());
+        let measurement = ClassicalRegister {
+            constant_circuit: None,
+            circuits: vec![circuit.clone()],
+        };
+        let program = QuantumProgram::ClassicalRegister {
+            measurement,
+            input_parameter_names: vec![],
+        };
+        let job_loc = api_backend_new.post_job(program);
+        assert!(job_loc.is_err());
 
-    let job_loc_dummy: String = "DummyString".to_string();
-    let job_status = api_backend_new.get_job_status(job_loc_dummy.clone());
-    assert!(job_status.is_err());
+        let job_loc_dummy: String = "DummyString".to_string();
+        let job_status = api_backend_new.get_job_status(job_loc_dummy.clone());
+        assert!(job_status.is_err());
 
-    let job_result = api_backend_new.get_job_result(job_loc_dummy.clone());
-    assert!(job_result.is_err());
+        let job_result = api_backend_new.get_job_result(job_loc_dummy.clone());
+        assert!(job_result.is_err());
 
-    let job_delete = api_backend_new.delete_job(job_loc_dummy);
-    assert!(job_delete.is_err());
+        let job_delete = api_backend_new.delete_job(job_loc_dummy);
+        assert!(job_delete.is_err());
+    }
 }
 
 /// Test error cases. Case 4: invalid job_id
